requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
asyncpg>=0.29.0
databases[postgresql]>=0.8.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Report generation
matplotlib>=3.7.0
weasyprint>=60.0
jinja2>=3.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.0.0
flake8>=6.0.0

# Optional performance enhancements
# cupy-cuda12x>=12.0.0  # Uncomment for GPU acceleration
# cython>=3.0.0         # Uncomment for compiled extensions 